Infraestructura/DataAnalizer/requirements.txt
tornado==6.1
pymongo==3.12.1
python-dateutil==2.8.2
PyJWT==2.3.0
tweepy==4.13.0
spacy==3.2.0
nltk==3.6.5
gender-guesser==0.4.0
country-list==0.1.3
pandas==1.3.4
tensorflow==2.7.0
tweepy==4.13.0
gender-guesser==0.4.0
country-list==0.1.3